+++ handler.go
@@ -2,8 +2,8 @@ package healthz
 
 import "net/http"
 
-// NewHealthServiceHandler creates an http.Handler from user configured Checkers
-// The handler is a standard http.ServeMux
+// NewHealthServiceHandler creates an http.Handler from user configured Checkers.
+// The returned handler is a standard http.ServeMux
 func NewHealthServiceHandler(livenessChecker Checker, readinessChecker Checker) http.Handler {
 	healthService := NewHealthService(livenessChecker, readinessChecker)
 	mux := http.NewServeMux()
@@ -14,12 +14,18 @@ func NewHealthServiceHandler(livenessChecker Checker, readinessChecker Checker) 
 	return mux
 }
 
-// HealthStatus checks if the application is healthy
+// HealthStatus checks if the application is healthy.
+//
+// This is identical to liveness checks in Kubernetes.
+// These checks are usually responsible for checking if the app is running (eg. listens on a specific port).
 func (s *HealthService) HealthStatus(w http.ResponseWriter, r *http.Request) {
 	s.checkStatus(w, r, s.livenessChecker)
 }
 
-// ReadinessStatus checks if the app is ready for accepting request (eg. database is available as well)
+// ReadinessStatus checks if the app is ready for accepting request.
+//
+// This is identical to readiness checks in Kubernetes.
+// These checks are usually responsible for checking if the app is functional (eg. database is available).
 func (s *HealthService) ReadinessStatus(w http.ResponseWriter, r *http.Request) {
 	s.checkStatus(w, r, s.readinessChecker)
 }