+++ testing.go
@@ -1,17 +1,17 @@
 package healthz
 
-// AlwaysSuccessChecker always returns success as the check result
+// AlwaysSuccessChecker always returns success as the check result.
 type AlwaysSuccessChecker struct{}
 
-// Check implements the Checker interface
+// Check implements the Checker interface.
 func (c *AlwaysSuccessChecker) Check() error {
 	return nil
 }
 
-// AlwaysFailureChecker always returns failure as the check result
+// AlwaysFailureChecker always returns failure as the check result.
 type AlwaysFailureChecker struct{}
 
-// Check implements the Checker interface
+// Check implements the Checker interface.
 func (c *AlwaysFailureChecker) Check() error {
 	return ErrCheckFailed
 }