+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - HTTP Timeout option
 - TCP timeout option
 
+### Changed
+
+- Improved documentation
+
 
 ## 0.5.0 - 2017-03-09